binance_to_cs/requirements.txt

asyncio
python-dotenv
cryptography
requests
websockets